Added handling more than 1 dataset. Few minor bugs still exist
