+++ lineRegressionGFK/Converters/PointYScalePositionConverter.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Data;
+
+namespace lineRegressionGFK.Converters
+{
+    public class PointYScalePositionConverter : IMultiValueConverter
+    {
+        public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (values.Length == 3)
+            {
+                var y = (double)values[0];
+                var scale = (double)values[1];
+                var pointRadius = (int)values[2];
+                return -y * scale - pointRadius / 2.0 + 15;
+            }
+            else
+                return values[0];
+        }
+
+        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
